fix: load optimization
